Scala.JS support
