+++ zio-json/js/src/main/scala/zio/JsonPackagePlatformSpecific.scala
@@ -1,0 +1,3 @@
+package zio
+
+trait JsonPackagePlatformSpecific {}
