+++ zio-json/js/src/main/scala/zio/json/JsonDecoderPlatformSpecific.scala
@@ -1,0 +1,3 @@
+package zio.json
+
+trait JsonDecoderPlatformSpecific[A] { self: JsonDecoder[A] => }
