+++ zio-json/js/src/main/scala/zio/json/JsonEncoderPlatformSpecific.scala
@@ -1,0 +1,3 @@
+package zio.json
+
+trait JsonEncoderPlatformSpecific[A] { self: JsonEncoder[A] => }
